@@ -1,6 +1,6 @@
 {
   "name": "reindex",
-  "version": "0.0.0",
+  "version": "2.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --port 6200",
@@ -8,52 +8,53 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "clean": "rimraf node_modules"
+    "clean": "rimraf node_modules",
+    "npm_install": "npm install --ignore-script --force"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.0.0",
-    "@angular/cdk": "^7.0.0",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
-    "@angular/forms": "~7.0.0",
-    "@angular/http": "~7.0.0",
-    "@angular/material": "^7.0.0",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
+    "@angular/animations": "^15.2.3",
+    "@angular/cdk": "^15.2.3",
+    "@angular/common": "^15.2.3",
+    "@angular/compiler": "^15.2.3",
+    "@angular/core": "^15.2.3",
+    "@angular/forms": "^15.2.3",
+    "@angular/http": "^7.0.0",
+    "@angular/material": "^15.2.3",
+    "@angular/platform-browser": "^15.2.3",
+    "@angular/platform-browser-dynamic": "^15.2.3",
+    "@angular/router": "^15.2.3",
     "angular2-prettyjson": "^3.0.1",
     "core-js": "^2.5.4",
-    "hammerjs": "^2.0.8",
     "http-server": "^0.12.3",
     "ngx-json-viewer": "^2.4.0",
-    "ngx-toastr": "^10.0.0",
-    "node-sass": "^4.14.1",
-    "rxjs": "~6.3.3",
+    "ngx-toastr": "^15.2.0",
+    "rxjs": "~6.6.7",
     "subsink": "^1.0.1",
+    "tslib": "^2.0.0",
     "yamljs": "^0.3.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.7",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "^15.2.4",
+    "@angular/cli": "^15.2.4",
+    "@angular/compiler-cli": "^15.2.3",
+    "@angular/language-service": "^15.2.3",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "node-sass": "^8.0.0",
+    "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "tslint": "~6.1.0",
+    "typescript": "~4.9.5"
   }
 }
