@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { FailuresDialogComponent } from './failures-dialog.component';
 
@@ -6,7 +6,7 @@ describe('FailuresDialogComponent', () => {
   let component: FailuresDialogComponent;
   let fixture: ComponentFixture<FailuresDialogComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FailuresDialogComponent ]
     })
