@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ReindexComponent } from './reindex.component';
 
@@ -6,7 +6,7 @@ describe('ReindexComponent', () => {
   let component: ReindexComponent;
   let fixture: ComponentFixture<ReindexComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ReindexComponent ]
     })
