@@ -25,7 +25,8 @@ import {AppConfigService} from './services/app-config.service';
 import {AppConfig} from './models/app.config';
 import {ErrorComponent} from './dialogs/error/error.component';
 import { DisableEnterDirective } from './directives/disable-enter.directive';
-import {MatSortModule} from '@angular/material';
+import { MatSortModule } from '@angular/material/sort';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 
 export function initializeAppFn(appConfigService: AppConfigService) {
   return () => {
@@ -40,24 +41,24 @@ export function initializeAppFn(appConfigService: AppConfigService) {
 // }
 
 @NgModule({
-  declarations: [
-    ReindexComponent,
-    YesNoDialogComponent,
-    HeaderComponent,
-    MainComponent,
-    SavedProjectsComponent,
-    ProjectMonitoringComponent,
-    CustomSpinnerDirective,
-    SearchInSourceIndexListPipe,
-    ReportDialogComponent,
-    ProjectComponent,
-    ViewDialogComponent,
-    PrettyJsonPipe,
-    ProgressComponent,
-    FailuresDialogComponent,
-    ErrorComponent,
-    DisableEnterDirective,
-  ],
+    declarations: [
+        ReindexComponent,
+        YesNoDialogComponent,
+        HeaderComponent,
+        MainComponent,
+        SavedProjectsComponent,
+        ProjectMonitoringComponent,
+        CustomSpinnerDirective,
+        SearchInSourceIndexListPipe,
+        ReportDialogComponent,
+        ProjectComponent,
+        ViewDialogComponent,
+        PrettyJsonPipe,
+        ProgressComponent,
+        FailuresDialogComponent,
+        ErrorComponent,
+        DisableEnterDirective,
+    ],
     imports: [
         CommonModule,
         MaterialModule,
@@ -69,30 +70,28 @@ export function initializeAppFn(appConfigService: AppConfigService) {
         NgxJsonViewerModule,
         PrettyJsonModule,
     ],
-  exports: [
-    ReindexComponent,
-  ],
-  providers: [
-    {
-      provide: AppConfig,
-      deps: [HttpClient],
-      useExisting: AppConfigService
-    },
-    {
-      provide: APP_INITIALIZER,
-      multi: true,
-      deps: [AppConfigService],
-      useFactory: initializeAppFn
-    },
-    // {
-    //   provide: APP_INITIALIZER,
-    //   multi: true,
-    //   deps: [AppConfigService],
-    //   useFactory: getUrl
-    // }
-  ],
-
-  entryComponents: [YesNoDialogComponent]
+    exports: [
+        ReindexComponent,
+    ],
+    providers: [
+        {
+            provide: AppConfig,
+            deps: [HttpClient],
+            useExisting: AppConfigService
+        },
+        {
+            provide: APP_INITIALIZER,
+            multi: true,
+            deps: [AppConfigService],
+            useFactory: initializeAppFn
+        },
+        // {
+        //   provide: APP_INITIALIZER,
+        //   multi: true,
+        //   deps: [AppConfigService],
+        //   useFactory: getUrl
+        // }
+    ]
 })
 export class ReindexModule {
 }
